Fix the graph.
Tried to make some changes here. We now do the following.

Color now follows energy use. If it takes a lot of energy, it's more red. If it
gives a lot of energy, it is more green.

Size now is based on duration. The more time it uses the bigger the item.